+++ tests/test_document_close_clears_active_route.py
@@ -30,7 +30,7 @@ import routes.document_routes as droutes
 from core.database import Document
 from core.database import Session as DbSession
 from routes.document_helpers import DocumentPatch
-from src.tool_implementations import set_active_document, get_active_document
+from src.agent_tools.document_tools import set_active_document, get_active_document
 
 _TMPDB = tempfile.NamedTemporaryFile(suffix=".db", delete=False)
 _ENGINE = create_engine(
