+++ tests/test_owned_document_query.py
@@ -1,5 +1,5 @@
 """Tests for _owned_document_query owner scoping (src/tool_implementations.py)."""
-from src.tool_implementations import _owned_document_query
+from src.agent_tools.document_tools import _owned_document_query
 
 
 class _FakeQuery:
